Fix installation on AIX
There are two things needed to get a successful build on AIX using GCC.

The first is similar to Windows: override the generated "export Init_libyajl"
instructions, which are present because the Ruby mkmf.rb module that this is
reusing assumes you're building a Ruby module.

The second is a workaround for what looks like bad configuration in Ruby
itself. I avoided trying to fix this in Ruby because it looks like the
argument we need to pass to GCC at link time (-shared) was deliberately removed
in 2010 to fix some other problem. The problem it fixed isn't well described and
I didn't find any relevant info online, so I'm not confident proposing that Ruby
reverts that change.

This change was tested on powerpc-ibm-aix7.2.0.0 using GCC 6.3.0 and Ruby 2.4.0.

Signed-off-by: Sam Thursfield <sam.thursfield@codethink.co.uk>